+++ examples/CCDCViz.py
@@ -1,5 +1,5 @@
 """
-   Copyright 2021 Ian Housman
+   Copyright 2022 Ian Housman
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -31,8 +31,8 @@ Map.clearMap()
 ####################################################################################################
 #Bring in ccdc image asset
 #This is assumed to be an image of arrays that is returned from the ee.Algorithms.TemporalSegmentation.Ccdc method
-ccdcImg = ee.ImageCollection("projects/CCDC/USA_V2")\
-          .filter(ee.Filter.eq('spectral', 'SR')).mosaic()
+ccdcImg = ee.ImageCollection('projects/lcms-292214/assets/CONUS-LCMS/Base-Learners/CCDC-Collection-1984-2021')\
+          .select(['tStart','tEnd','tBreak','changeProb','red.*','nir.*','swir1.*','swir2.*','NDVI.*','NBR.*']).mosaic()
 # ccdcImg = ee.ImageCollection('projects/lcms-292214/assets/R8/PR_USVI/Base-Learners/CCDC-Landsat_1984_2020').mosaic()
 #Specify which harmonics to use when predicting the CCDC model
 #CCDC exports the first 3 harmonics (1 cycle/yr, 2 cycles/yr, and 3 cycles/yr)
@@ -46,6 +46,10 @@ fillGaps = False
 #Specify which band to use for loss and gain. 
 #This is most important for the loss and gain magnitude since the year of change will be the same for all years
 changeDetectionBandName = 'NDVI'
+
+
+# Choose whether to show the most recent ('mostRecent') or highest magnitude ('highestMag') CCDC break
+sortingMethod = 'mostRecent'
 ####################################################################################################
 #Pull out some info about the ccdc image
 startJulian = 1
@@ -54,14 +58,16 @@ startYear = 1984
 endYear = 2020
 
 #Add the raw array image
-Map.addLayer(ccdcImg,{},'Raw CCDC Output',False)
+Map.addLayer(ccdcImg,{'opacity':0},'Raw CCDC Output',False)
 
 #Extract the change years and magnitude
 changeObj = changeDetectionLib.ccdcChangeDetection(ccdcImg,changeDetectionBandName);
-Map.addLayer(changeObj['highestMag']['loss']['year'],{'min':startYear,'max':endYear,'palette':changeDetectionLib.lossYearPalette},'Loss Year')
-Map.addLayer(changeObj['highestMag']['loss']['mag'],{'min':-0.5,'max':-0.1,'palette':changeDetectionLib.lossMagPalette},'Loss Mag',False);
-Map.addLayer(changeObj['highestMag']['gain']['year'],{'min':startYear,'max':endYear,'palette':changeDetectionLib.gainYearPalette},'Gain Year');
-Map.addLayer(changeObj['highestMag']['gain']['mag'],{'min':0.05,'max':0.2,'palette':changeDetectionLib.gainMagPalette},'Gain Mag',False);
+changeDetectionLib.lossMagPalette = changeDetectionLib.lossMagPalette.split(',')
+changeDetectionLib.lossMagPalette.reverse()
+Map.addLayer(changeObj[sortingMethod]['loss']['year'],{'min':startYear,'max':endYear,'palette':changeDetectionLib.lossYearPalette},'Loss Year')
+Map.addLayer(changeObj[sortingMethod]['loss']['mag'],{'min':-0.5,'max':-0.1,'palette':changeDetectionLib.lossMagPalette},'Loss Mag',False);
+Map.addLayer(changeObj[sortingMethod]['gain']['year'],{'min':startYear,'max':endYear,'palette':changeDetectionLib.gainYearPalette},'Gain Year');
+Map.addLayer(changeObj[sortingMethod]['gain']['mag'],{'min':0.05,'max':0.2,'palette':changeDetectionLib.gainMagPalette},'Gain Mag',False);
 
 #Apply the CCDC harmonic model across a time series
 #First get a time series of time images 
@@ -70,6 +76,7 @@ yearImages = changeDetectionLib.getTimeImageCollection(startYear,endYear,startJu
 #Then predict the CCDC models
 fitted = changeDetectionLib.predictCCDC(ccdcImg,yearImages,fillGaps,whichHarmonics)
 Map.addLayer(fitted.select(['.*_predicted']),{'opacity':0},'Fitted CCDC',True);
+Map.addLayer(fitted.filter(ee.Filter.calendarRange(1990,1990,'year')).select(['.*_predicted']),{'opacity':0},'Fitted CCDC 1990',True);
 
 ####################################################################################################
 #Load the study region
@@ -77,4 +84,5 @@ studyArea = ccdcImg.geometry()
 Map.addLayer(studyArea, {'strokeColor': '0000FF'}, "Study Area", False)
 # Map.centerObject(studyArea)
 ####################################################################################################
+Map.turnOnInspector()
 Map.view()