+++ examples/LANDTRENDRWrapper.py
@@ -1,5 +1,5 @@
 """
-   Copyright 2021 Ian Housman
+   Copyright 2022 Ian Housman
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -26,6 +26,7 @@ sys.path.append(os.getcwd())
 #Module imports
 import geeViz.getImagesLib as getImagesLib
 import geeViz.changeDetectionLib as changeDetectionLib
+import geeViz.taskManagerLib as taskManagerLib
 ee = getImagesLib.ee
 Map = getImagesLib.Map
 Map.clearMap()
@@ -109,7 +110,7 @@ outputName = 'LT_Test'
 
 #Provide location composites will be exported to
 #This should be an asset folder, or more ideally, an asset imageCollection
-exportPathRoot = 'users/iwhousman/test/ChangeCollection'
+exportPathRoot = 'users/username/someCollection'
 
 
 
@@ -172,4 +173,11 @@ if exportLTStack:
 Map.addLayer(studyArea, {'strokeColor': '0000FF'}, "Study Area", False)
 Map.centerObject(studyArea)
 ####################################################################################################
+####################################################################################################
+# View map
+Map.turnOnInspector()
 Map.view()
+####################################################################################################
+####################################################################################################
+# If exporting LT stack, track the exports
+if exportLTStack:taskManagerLib.trackTasks2()