+++ examples/getCombinedLandsatSentinel2Wrapper.py
@@ -1,5 +1,5 @@
 """
-   Copyright 2021 Ian Housman
+   Copyright 2022 Ian Housman
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -22,6 +22,7 @@ sys.path.append(os.getcwd())
 
 #Module imports
 import geeViz.getImagesLib as getImagesLib
+import geeViz.taskManagerLib as taskManagerLib
 ee = getImagesLib.ee
 Map = getImagesLib.Map
 Map.clearMap()
@@ -229,7 +230,7 @@ outputName = 'Landsat_Sentinel2_Hybrid'
 
 # Provide location composites will be exported to
 # This should be an asset folder, or more ideally, an asset imageCollection
-exportPathRoot = 'users/iwhousman/test/compositeCollection'
+exportPathRoot = 'users/username/someCollection'
 
 
 
@@ -275,4 +276,10 @@ Map.addLayer(studyArea, {'strokeColor': '0000FF'}, "Study Area", True)
 Map.centerObject(studyArea)
 ####################################################################################################
 ####################################################################################################
+# View map
+Map.turnOnInspector()
 Map.view()
+####################################################################################################
+####################################################################################################
+# If exporting composites, track the exports
+if exportComposites:taskManagerLib.trackTasks2()