+++ examples/lcmsViewerExample.py
@@ -1,6 +1,6 @@
 '''
 Based on https://github.com/google/earthengine-community/blob/master/datasets/scripts/LCMS_Visualization.js
- * Copyright 2021 The Google Earth Engine Community Authors
+ * Copyright 2022 The Google Earth Engine Community Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.